refactor(frontend): split services into their respective domains
