+++ frontend/app/.server/domain/person-case/services/index.ts
@@ -1,0 +1,7 @@
+export * as applicantGenderService from './applicant-gender-service';
+export * as applicantPrimaryDocumentService from './applicant-primary-document-service';
+export * as applicantSecondaryDocumentService from './applicant-secondary-document-service';
+export * as applicantSinService from './applicant-sin-service';
+export * as applicationSubmissionScenario from './application-submission-scenario';
+export * as applicationTypeService from './application-type-service';
+export * as languageCorrespondenceService from './language-correspondence-service';
