+++ frontend/app/routes/protected/person-case/parent-details.tsx
@@ -11,12 +11,7 @@ import * as v from 'valibot';
 import type { Info, Route } from './+types/parent-details';
 
 import { serverEnvironment } from '~/.server/environment';
-import {
-  getCountries,
-  getLocalizedCountries,
-  getLocalizedProvinces,
-  getProvinces,
-} from '~/.server/services/locale-data-service';
+import { countryService, provinceService } from '~/.server/shared/services';
 import { requireAuth } from '~/.server/utils/auth-utils';
 import { i18nRedirect } from '~/.server/utils/route-utils';
 import { Button } from '~/components/button';
@@ -49,8 +44,8 @@ export async function loader({ context, request }: Route.LoaderArgs) {
 
   return {
     documentTitle: t('protected:parent-details.page-title'),
-    localizedCountries: getLocalizedCountries(lang),
-    localizedProvincesTerritoriesStates: getLocalizedProvinces(lang),
+    localizedCountries: countryService.getLocalizedCountries(lang),
+    localizedProvincesTerritoriesStates: provinceService.getLocalizedProvinces(lang),
     PP_CANADA_COUNTRY_CODE,
     defaultFormValues: sessionData.map((details) =>
       details.unavailable
@@ -122,7 +117,7 @@ export async function action({ context, request }: Route.ActionArgs) {
                     v.object({
                       country: v.literal(PP_CANADA_COUNTRY_CODE, t('protected:parent-details.country-error.invalid-country')),
                       province: v.picklist(
-                        getProvinces().map(({ id }) => id),
+                        provinceService.getProvinces().map(({ id }) => id),
                         t('protected:parent-details.province-error.required-province'),
                       ),
                       city: v.pipe(
@@ -139,7 +134,7 @@ export async function action({ context, request }: Route.ActionArgs) {
                         v.nonEmpty(t('protected:parent-details.country-error.required-country')),
                         v.excludes(PP_CANADA_COUNTRY_CODE, t('protected:parent-details.country-error.invalid-country')),
                         v.picklist(
-                          getCountries().map(({ id }) => id),
+                          countryService.getCountries().map(({ id }) => id),
                           t('protected:parent-details.country-error.invalid-country'),
                         ),
                       ),
